Refactor AddProjectSection component, add project details functionality, and update profile section
